server/requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
flask-cors
python-dotenv
groq
gtts
chromadb
requests
moviepy
Pillow
numpy
psycopg2-binary
PyJWT==2.8.1
bcrypt==4.1.2
Flask-JWT-Extended==4.5.3
